Add files via upload
Simple ATM System project is written in Python. The project file contains a python script (simpleatm.py). This is a simple console based system which is very easy to use. Talking about the system, it contains limited functions which only includes Withdrawing and Depositing amount. Here, at first the user has to enter existing account holder name, when the name matches the system proceed toward the next procedure i.e asking pin number. When a user passes all these sign-in procedures, he/she can use the features. It is too easy to use, he/she just has to enter the amount whether in deposit section or the withdraw section. The system calculates the total remaining balance of the 

respective account and displays to the user.

This simple console based ATM system provides the simple account balance management of a respective account. It contains fewer features but the essential one. There is no database connection or neither any external text or other files used in this mini project to save user’s data. After each transaction using the system, the system redirects the user to the login page and he/she can log in again with the other account too.

In order to run the project, you must have installed Python, on your PC. This is a simple Console Based system, specially written for the beginners. Simple ATM System in Python project with source code is free to download. Use for education purpose only! For the project demo, have a look at the image slider below.