Description
Ludo Game The Ludo Game with Source Code is a multi-player platform game where the players must race against each other to reach the home triangle. The project purpose is to have fun leisure time and enjoy your bonding with family, friends, school classmate, or co-worker. The project was built in a simple console application. The project is played in a special board called "Ludo Board" a board with a square pattern with a shape of cross, each arm is divided into three adjacent columns of eight squares. The players can play single or multi gameplay depend on what you choose in the option, in single gameplay you can add bot players either duo, trio, or quadro. While in multi-player you can add a player maximum of 4 with name on it to identify the player position. The game is so simple to play with, to move your piece you need to roll a six-sided die. Each time you a roll six-sided die you can move a new piece from the starting circle into the first square of your track. This project is a simple game that is similar to a popular game called "Ludo King" where your main goal is to move all your pieces to your home triangle. The game is very fun to play with friends, race your 4 pieces to move to the home triangle in order to beat your opponent. The Ludo Game is a very addictive board style game where players goal is to race against each other.