Description
Ball Breaker Game Ball Breaker Game with Source Code is a Python program that is a single-player where the players must break all the balls from the opposite side. The project is a simple game that can be enjoy of any person that needed some past time. The project was created by the use of Pygame module where it contains a different platform that use to developed a game. The project is played in a simple black background where the top side contains a different colored ball while the bottom is only a blue horizontal platform. The player can move the platform via keyboard (Left Arrow to move left, Right Arrow to move right). The player goal is to clear all the colored balls above using the moving white ball. The game will be over if the player failed to hit the moving ball. The Ball Breaker Game is very enjoyable and fun to play with, try to clear all the colored balls and achieve a higher score.