Description
Cafe Billing System with Source Code is a Python program that can compute the total bill of a customer digitally. This program was built using only Python language. The system contains a simple GUI application that use tkinter framework. The user can only enter the quantity of an items ordered by the customers. The system can display the list of item price when clicking the PRICE button, and also when the item is ordered it generate the total bill of the purchase item by clicking the TOTAL button. The order no is auto generated after purchase.